backend/pyproject.toml
[tool.aerich]
tortoise_orm = "config.db.tortoise_config"
location = "./migrations"
src_folder = "./."

[tool.poetry]
name = "api"
version = "0.1.0"
description = "API untuk project myPUHH"
authors = ["manweljs <manweljs@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
aerich = "^0.7.2"
asyncpg = "^0.29.0"
fastapi = "^0.111.0"
tortoise-orm = "^0.20.1"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.1.3"
pytz = "^2024.1"
fastapi-crudrouter = "^0.8.6"
pytest = "^8.2.0"
httpx = "^0.27.0"
pandas = "^2.2.2"
boto3 = "1.34.69"
botocore = "1.34.69"
aiobotocore = "2.12.3"
fastapi-pagination = "^0.12.24"
aiofiles = "^23.2.1"
openpyxl = "^3.1.2"

[tool.poetry.scripts]
dev = "api.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

